+++ backend/curriculum_tracking/tests/test_activity_log_entry_creators_envoked.py
@@ -1,14 +1,25 @@
 """for each of the log entry creators in curriculum_tracking.activity_log_entry_creators, make sure it is called when it should be and creates the correct log entries
 """
-from rest_framework.test import APITestCase
 
-from test_mixins import APITestCaseMixin
 from . import factories
-from core.tests.factories import UserFactory
-from curriculum_tracking.models import AgileCard, ContentItem
+from curriculum_tracking.models import (
+    AgileCard,
+    ContentItem,
+    RecruitProjectReview,
+    TopicReview
+)
+from curriculum_tracking.constants import NOT_YET_COMPETENT
 from activity_log.models import LogEntry
 from curriculum_tracking import activity_log_entry_creators as creators
 from curriculum_tracking.constants import COMPETENT
+from test_mixins import APITestCase, APITestCaseMixin
+from .factories import (
+    UserFactory,
+    AgileCardFactory,
+    RecruitProjectFactory,
+)
+from curriculum_tracking.tests.factories import ProjectContentItemFactory, ContentItemFactory, TopicProgressFactory
+import mock
 
 
 class log_card_started_Tests(APITestCase, APITestCaseMixin):
@@ -106,3 +117,107 @@ class log_project_competence_review_done_Tests(APITestCase, APITestCaseMixin):
             start_url, data={"status": COMPETENT, "comments": "blah more stuff"}
         )
         self.assertEqual(LogEntry.objects.count(), 2)
+
+
+class log_project_and_topic_competence_reviews_done_TESTS(APITestCase, APITestCaseMixin):
+    LIST_URL_NAME = "agilecard-list"
+    SUPPRESS_TEST_POST_TO_CREATE = True
+    FIELDS_THAT_CAN_BE_FALSEY = [
+        "code_review_competent_since_last_review_request",
+        "code_review_excellent_since_last_review_request",
+        "code_review_red_flag_since_last_review_request",
+        "code_review_ny_competent_since_last_review_request",
+        "requires_cards",
+        "required_by_cards",
+        "project_submission_type_nice",
+        "topic_needs_review",
+        "topic_progress",
+        "due_time",
+        "complete_time",
+        "review_request_time",
+        "start_time",
+        "tag_names",
+        "can_start",
+        "can_force_start",
+        "flavour_names",
+        "open_pr_count",
+        "oldest_open_pr_updated_time",
+        "users_that_reviewed_since_last_review_request"
+    ]
+
+    def verbose_instance_factory(self):
+        project = RecruitProjectFactory()
+        card = AgileCardFactory(recruit_project=project)
+        return card
+
+    def setUp(self):
+        self.api_url = self.get_list_url()
+
+    @mock.patch('curriculum_tracking.activity_log_entry_creators.log_topic_competence_review_done')
+    @mock.patch('curriculum_tracking.activity_log_entry_creators.log_project_competence_review_done')
+    def test_log_project_competence_review_invoked_from_api_endpoint_for_project_review(
+            self, log_project_competence_review_done, log_topic_competence_review_done
+    ):
+
+        super_user = UserFactory(is_superuser=True)
+        card = AgileCardFactory(content_item=ProjectContentItemFactory())
+        self.login(super_user)
+        response = self.client.post(
+                path=f'{self.get_instance_url(pk=card.id)}add_review/',
+                data={"status": NOT_YET_COMPETENT, "comments": "dammit"}
+            )
+        self.assertTrue(response.status_code, 200)
+        project_review = RecruitProjectReview.objects.first()
+        log_project_competence_review_done.assert_called_with(project_review)
+        log_topic_competence_review_done.assert_not_called()
+
+    @mock.patch('curriculum_tracking.activity_log_entry_creators.log_topic_competence_review_done')
+    @mock.patch('curriculum_tracking.activity_log_entry_creators.log_project_competence_review_done')
+    def test_log_topic_competence_review_invoked_from_api_endpoint_for_topic_review(
+            self, log_project_competence_review_done, log_topic_competence_review_done
+    ):
+        super_user = UserFactory(is_superuser=True)
+        card = AgileCardFactory(content_item=ContentItemFactory())
+        card.topic_progress = TopicProgressFactory()
+        card.save()
+        self.login(super_user)
+        response = self.client.post(
+            path=f'{self.get_instance_url(pk=card.id)}add_review/',
+            data={"status": NOT_YET_COMPETENT, "comments": "dammit"}
+        )
+        topic_review = TopicReview.objects.first()
+        self.assertTrue(response.status_code, 200)
+        log_topic_competence_review_done.assert_called_with(topic_review)
+        log_project_competence_review_done.assert_not_called()
+
+    def test_log_project_competence_review_invoked_creates_log_entries(self):
+
+        super_user = UserFactory(is_superuser=True)
+        card = AgileCardFactory(content_item=ProjectContentItemFactory())
+        self.login(super_user)
+        response = self.client.post(
+            path=f'{self.get_instance_url(pk=card.id)}add_review/',
+            data={"status": NOT_YET_COMPETENT, "comments": "dammit"}
+        )
+        self.assertTrue(response.status_code, 200)
+        project_review = RecruitProjectReview.objects.first()
+        log_entry = LogEntry.objects.first()
+        self.assertEqual(log_entry.actor_user, project_review.reviewer_user)
+        self.assertEqual(log_entry.timestamp, project_review.timestamp)
+
+    def test_log_topic_competence_review_invoked_creates_log_entries(self):
+
+        super_user = UserFactory(is_superuser=True)
+        card = AgileCardFactory(content_item=ContentItemFactory())
+        card.topic_progress = TopicProgressFactory()
+        card.save()
+        self.login(super_user)
+        response = self.client.post(
+            path=f'{self.get_instance_url(pk=card.id)}add_review/',
+            data={"status": NOT_YET_COMPETENT, "comments": "dammit"}
+        )
+        self.assertTrue(response.status_code, 200)
+        topic_review = TopicReview.objects.first()
+        log_entry = LogEntry.objects.first()
+        self.assertEqual(log_entry.actor_user, topic_review.reviewer_user)
+        self.assertEqual(log_entry.timestamp, topic_review.timestamp)
